Removed pointers between dungeon tiles.
Removed the north, south, east and west pointers in struct Tile.  Dungeon
tiles no longer form a graph, as the graph was mostly not used for finding
tiles, but instead the struct Tiles collection is searched.

Removed functions that manipulated the Tile graph, including
destroyTileGraph() and visitTiles().  Changed assertion in DungeonTest that
used the north Tile struct member.

Fixed analyzer problems in MagicItem.c related to lack of defaults in two
switch statements.