Added tilesOnLevel() function to select a Tiles subset.
The tilesOnLevel() function returns a const struct Tiles that refers to a
subset of a Tiles struct where all tiles have a given z coordinate (level).
This initial implementation is a simple linear search, will replace it with
a binary lower/upper bound search in the future.

Added unit tests for tilesOnLevel().

Renamed function gatherDungeonStatistics() to gatherTileStatistics() and
moved it and supporting functions from Dungeon.h/Dungeon.c to Tiles.h/Tiles.c.

Renamed struct DungeonStatistics to TileStatistics.

Renamed graphDungeonUsingText() to graphDungeonLevelUsingText() and added a
z (level) parameter.  Changed the implementation to filter the dungeon tiles
using tilesOnLevel().  Similarly, changed graphDungeonUsingText() to
graphDungeonLevelUsingText().

Changed the signature of findTileInTilesAt() to take a const struct Tiles *
so it plays nicely with const Tiles collections.