Added ranges for x, y and z coordinates to Tiles struct.
Functions xRangeOfTiles(), yRangeOfTiles() and zRangeOfTiles() return ranges
for the respective coordinate for tiles in the Tiles struct.  Added
corresponding range members to struct Tiles.  Modified addTileToTiles() and
removeTileFromTiles() functions to update the tile ranges.

Changed graphDungeonLevelUsingText() to use ranges from struct Tiles instead
of calling gatherTileStatistics().

Added a bunch of useful functions for Range and ReverseRange structs:
equalRanges() and equalReverseRanges() for comparing two Range/ReverseRange
structs for equality; extendRangeToIncludeValue() and
extendReverseRangeToIncludeValue() for extendint a Range/ReverseRange;
isEmptyRange() and isEmptyReverseRange() for testing for zero length;
lengthOfRange() and lengthOfReverseRange() for calculating the length.

Added RangeTest.c with unit tests for Range and ReverseRange functions.