Add precheck script for `check` target.
The precheck script removes code coverage files before the executables
in the `check` target are run.  This fixes an intermittent problem where
errors occurred when running `check` repeatedly and the coverage library
attempts to merge in new coverage data with the old coverage data.

Also added the `-g` and `-O0` flags to the build.