Added first "tiles thumbnail" function to streamline testing.
The function `tiles_thumbnail_types_alloc()` creates a text rectangle
with a simple representation of an array of tiles, like this:

            "   -2 -1  0  1 \n"
            " 1  :  :  :  : \n"
            " 0  :  ^  ^  : \n"
            "-1     :  :  : \n";

The thumbnail is easy to create, and can be checked against a generated
array of tile structs using `str_eq()`, replacing many lines of asserts.

Implemented `tiles_thumbnail_types_alloc()` and used it in
`digger_dig_starting_stairs_test()` and `digger_dig_area_test()` in
addition to the existing asserts.  Added unit tests for
`tiles_thumbnail_types_alloc()` as well.