Set `CMAKE_OSX_SYSROOT` to fix local build problem.
Set the `CMAKE_OSX_SYSROOT` property to the generic `SDKs/MacOSX.sdk`
path in Xcode instead of the `SDKs/MacOSX10.14.sdk` path to enable
building with the combination of Xcode 11 and macOS 10.14 Mojave.

Xcode 11 only ships with the SDK for macOS 10.15 Catalina, but CMake
detects the OS version and looks for the matching SDK path in Xcode.