Fix broken test `generator_delete_digger_test()`.
Change the behavior of `realloc_or_die()` to ensure that it always
returns a non-NULL pointer.  On macOS, `realloc()` always returns
non-NULL, even when the new size is zero.  On Linux, reallocating to
zero size will return NULL.  Simulate the macOS behavior by checking for
this and allocating a new item of size one.