Add `HOMEBREW_NCURSES` build option.
Work around broken ncurses `libmenu` on macOS by adding the build option
`HOMEBREW_NCURSES`, that will look for the Homebrew version of ncurses
installed in /usr/local/opt/ncurses.

Update the Travis build to set `HOMEBREW_NCURSES` on for macOS builds.
Update the readme with a better description of the link error you get
with the broken `libmenu` and command line to enable `HOMEBREW_NCURSES`.