Add boolean build options `WALL` and `ADDRESS_SANITIZER`.
When the `WALL` option is `ON`, add the `-Wall -Werror` flags to the
global compile options.

When the `ADDRESS_SANITIZER` option is `ON`, add the
`-fsanitize=address` flag to global compile and link options.

Update the logic for the `COVERAGE` option to use `add_link_options()`
instead of `set(CMAKE_EXE_LINKER_FLAGS)`.

Update the Sourcehut Debian build to enable build options
`ADDRESS_SANITIZER`, `COVERAGE` and `WALL`.  Also add a `coverage` build
task to send coverage data to codecov.io.

Remove the Travis build badge from the readme file.