Revamp the readme file.
Add some details about the `fnf` and `fiends` executables and the
various functionality available.  Add links for `tvision`.  Add info on
the `GRAPH_TARGETS` build option.  Drop the "Motivation" paragraph.

Update the copyright year in the license file and add the missing
`tvision` copyright file.

Set the C++ standard to C++ 14 for the Cmake build.