Pad lines in `TextRect` to a uniform width.
Change the `TextRect` constructors to pad each line with spaces to the
width of the longest line.  Also make the `appendLine()` and
`appendLines()` private methods into static helper functions.

Simplify the `draw()` methods of `TreasureTypeTableView` and
`TreasureView` a little since the lines in their source `TextRect` data
members are now uniform.