`TreasureView` holds on to the generated `struct treasure`.
Add a `unique_ptr` to the treasure to `TreasureView`.  Split the
`generateTreasure()` function into two parts: `generateTreasure()` now
returns the allocated and initialized `treasure` struct and
`treasureDetails()` turns the treasure into a `TextRect`.  Remove the
`const` qualifier from `TreausreView::text` so it can be assigned to in
the constructor body.

Added a move constructor and move assignment operator to `TextRect`.