Add the `HAS_REALLOCARRAY` feature macro.
Check for the existance of `reallocarray()` in the top-level CMakeLists
file.  Propogate the `HAS_REALLOCARRAY` variable to the `base` and
`base_tests` targets.

Also make sure to propogate `HAS_OPTRESET` to the `fnf` target.

Conditionally include the `reallocarray()` polyfill when
`HAS_REALLOCARRAY` is not defined.

Make sure `reallocarray()` sets `errno` to `ENOMEM` on overflow.  Don't
assume that `reallocarray()` must return a non-NULL pointer for a
zero-sized reallocation, since the same behavior in `realloc()` varies
on different platforms.